Logging in now works
Login by running the authentication server on ``cd backend`` and
then ``node server``. Then boot up the frontend by ``cd ~`` and
``npm run react-start``. Navigating to index.html gives you a form
in which you can enter in the username and password for an
already registered user. Submitting the form grants you a token,
that you may use to authenticate API requests in the future.